+++ CHANGELOG.md
@@ -7,3 +7,9 @@
     * create endpoint for getting sprites via proxy
     * create endpoint for getting random Pokemon spcies
 <img alt="pixelchu" src="scrots/pixelchu.PNG" width="250"/>
+
+# 2023-07-10
+
+* server
+    * create game registry to manage games
+    * setup websockets for duplex connection